partial completion of splitting ConnectionSession into two classes one for HTTP one for GRPC. Next: Fix unit test by splitting into two unit tests and having protocol specific unit tests and plan for protocol specific targets
